+++ frontend/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2017-04-01
+### Added
+- Add locale support for topbar menu.
+- Add model download in project settings.
+
+### Fixed
+
 ## 2017-03-30
 ### Added
 - Add application global fatal error message box with APP_FATAL_ERROR action.